backend/requirements.txt
blessed==1.19.1
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
coverage==6.4.2
Django==4.0.5
django-extensions==3.1.5
django-picklefield==3.1
django-q==1.3.9
django-rest-framework==0.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
PyJWT==2.3.0
python-dateutil==2.8.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1